aggressive refactoring
converted this from what seems like a purpose-built tool to a more
general unix-style utility, with argparse and stuff. defaults to reading
stdin and writing to stdout, or, if passed a file, reads that and spews
the results to stdout.

still need to hack in a file writing routine, but for now it does
what i want